updating payment-popup to recalculate boost rank on content change
