+++ pages/index.js
@@ -5,9 +5,9 @@ import Postmate from 'postmate';
 import PaymentPopup from '../components/payment-popup';
 import * as BoostHelpers from '../lib/boost-helpers';
 import * as ApiCompat from '../lib/api-compatibility';
-import * as LogSlider from '../components/payment-popup/log-slider';
 
 const DEFAULT_PROPS = ApiCompat.normalizeLegacyApi();
+
 
 const Home = () => {
 	const initialProps = DEFAULT_PROPS;
@@ -16,68 +16,13 @@ const Home = () => {
 	const [parent, setParent] = useState(null);
 	const [opened, setOpened] = useState(true);
 
-	const sliderMarkerStepsInit = (newProps, props)=>{
-		// Overrides min, max and initial when they are explicitly defined by the user
-		newProps.diff.min = 1;
-		// if (props.diff.min > 0) newProps.diff.min = 1; // props.diff.min;
-		if (props.diff.max > 0) newProps.diff.max = props.diff.max;
-		if (props.diff.initial > 0) newProps.diff.initial = props.diff.initial;
-
-		// Ensures safe initial value
-		if (newProps.diff.initial > newProps.diff.max) newProps.diff.initial = newProps.diff.max;
-		return newProps;
-	};
-
-	const updateBoostsRank = async props => {
-		if (!props.boostRank) {
-			return props;
-		}
-		// Fetches the API and rank calcs
-		let newProps = await BoostHelpers.prepareBoostProps(props);
-		
-		let contentBoosts = {};
-		
-		if (props.content.hash.length == 64) {
-			contentBoosts = await BoostHelpers.searchContentHex(props.content.hash, props);
-			const CBV = contentBoosts.totalDifficulty_;
-			props.contentBoosts = contentBoosts;
-
-			const ranksCtrl = LogSlider.GetTopNFromSignals(newProps.signals, CBV);
-			const newSliderCtrl = LogSlider.NewContentSliderCtrl(CBV, ranksCtrl, newProps);
-			// Overrides min, max and initial when they are explicitly defined by the user
-			newProps.diff.min = 1;
-			if (newSliderCtrl.MinBoost > 0) newProps.diff.min = newSliderCtrl.MinBoost;
-			if (newSliderCtrl.MaxBoost > 0) newProps.diff.max = newSliderCtrl.MaxBoost;
-			const toTop1 = newSliderCtrl.Top1Boost;
-			if (toTop1 > 0) newProps.diff.initial = toTop1;
-
-			// Ensures safe initial value
-			if (newProps.diff.initial > newProps.diff.max) newProps.diff.initial = newProps.diff.max;
-			
-			if (newProps.slider.rankMarkers === true || Array.isArray(newProps.slider.rankMarkers)) {
-				let rm = [1, 2, 3, 5, 10];
-				if (Array.isArray(newProps.slider.rankMarkers) && newProps.slider.rankMarkers.length > 0) {
-					rm = newProps.slider.rankMarkers;
-				}
-				newProps.slider.sliderRankMarkers = LogSlider.sliderRankMarkers(newSliderCtrl, rm);
-			} else {
-				newProps = sliderMarkerStepsInit(newProps, props);
-			}
-			newProps.sliderCtrl = newSliderCtrl;
-			
-		} else {
-			newProps = sliderMarkerStepsInit(newProps, props);
-		}
-		return newProps;
-	};
-
 	const startParentHandshake = async () => {
 		if (parent) return; // parent connects only once
 		try {
 			const parentHandshake = new Postmate.Model({
 				open: async userProps => {
 					let compatProps = ApiCompat.normalizeLegacyApi(userProps);
-					compatProps = await updateBoostsRank(compatProps);
+					compatProps = await BoostHelpers.updateBoostsRank(compatProps);
 					compatProps.opening = true;
 					setOpened(true);
 					setPaymentProps(compatProps);
